+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Full documentation for the ROCR Debug Agent library is available at
 [rocm.docs.amd.com/rocr_debug_agent](https://rocm.docs.amd.com/projects/rocr_debug_agent/en/latest/).
 
+## ROCR Debug Agent 2.2.0
+
+### Added
+- Add the `-t` and `--print-debugtrap` options to print waves when they
+  execute a `__builtin_debugtrap` instruction.
+
 ## ROCR Debug Agent 2.1.0 for ROCm 7.0
 
 ### Added